# Inventory loading