--- a/app/assets/javascripts/admin/index_utils/services/paged_fetcher.js.coffee
+++ b/app/assets/javascripts/admin/index_utils/services/paged_fetcher.js.coffee
@@ -2,19 +2,24 @@ angular.module("admin.indexUtils").factory "PagedFetcher", (dataFetcher) ->
   new class PagedFetcher
     # Given a URL like http://example.com/foo?page=::page::&per_page=20
     # And the response includes an attribute pages with the number of pages to fetch
-    # Fetch each page async, and call the processData callback with the resulting data
-    fetch: (url, processData, onLastPageComplete) ->
-      dataFetcher(@urlForPage(url, 1)).then (data) =>
-        processData data
+    # Fetch each page async, and call the pageCallback callback with the resulting data
+    # Developer note: this class should not be re-used!
+    page: 1
+    last_page: 1
 
-        if data.pages > 1
-          for page in [2..data.pages]
-            lastPromise = dataFetcher(@urlForPage(url, page)).then (data) ->
-              processData data
-          onLastPageComplete && lastPromise.then onLastPageComplete
-          return
-        else
-          onLastPageComplete && onLastPageComplete()
+    fetch: (url, pageCallback) ->
+      @fetchPages(url, @page, pageCallback)
 
     urlForPage: (url, page) ->
       url.replace("::page::", page)
+
+    fetchPages: (url, page, pageCallback) ->
+      dataFetcher(@urlForPage(url, page)).then (data) =>
+        @page++
+        @last_page = data.pages
+
+        pageCallback(data) if pageCallback
+
+        if @page <= @last_page
+          @fetchPages(url, @page, pageCallback)
+