--- a/app/controllers/api/products_controller.rb
+++ b/app/controllers/api/products_controller.rb
@@ -69,10 +69,10 @@ module Api
     end
 
     def overridable
-      producers = OpenFoodNetwork::Permissions.new(current_api_user).
-        variant_override_producers.by_name
+      producer_ids = OpenFoodNetwork::Permissions.new(current_api_user).
+        variant_override_producers.by_name.select('enterprises.id')
 
-      @products = paged_products_for_producers producers
+      @products = paged_products_for_producers producer_ids
 
       render_paged_products @products, ::Api::Admin::ProductSimpleSerializer
     end
@@ -118,11 +118,11 @@ module Api
       ]
     end
 
-    def paged_products_for_producers(producers)
+    def paged_products_for_producers(producer_ids)
       Spree::Product.scoped.
         merge(product_scope).
         includes(variants: [:product, :default_price, :stock_items]).
-        where(supplier_id: producers).
+        where(supplier_id: producer_ids).
         by_producer.by_name.
         ransack(params[:q]).result.
         page(params[:page]).per(params[:per_page])