--- a/app/models/spree/stock/quantifier.rb
+++ b/app/models/spree/stock/quantifier.rb
@@ -11,11 +11,7 @@ module Spree
       end
 
       def total_on_hand
-        if Spree::Config.track_inventory_levels
-          stock_items.sum(&:count_on_hand)
-        else
-          Float::INFINITY
-        end
+        stock_items.sum(&:count_on_hand)
       end
 
       def backorderable?