--- a/app/serializers/api/admin/product_simple_serializer.rb
+++ b/app/serializers/api/admin/product_simple_serializer.rb
@@ -1,0 +1,16 @@
+class Api::Admin::ProductSimpleSerializer < ActiveModel::Serializer
+  attributes :id, :name
+
+  has_one :supplier, key: :producer_id, embed: :id
+  has_many :variants, key: :variants, serializer: Api::Admin::VariantSimpleSerializer
+
+  def on_hand
+    return 0 if object.on_hand.nil?
+
+    object.on_hand
+  end
+
+  def price
+    object.price.nil? ? '0.0' : object.price
+  end
+end