--- a/app/serializers/api/admin/product_simple_serializer.rb
+++ b/app/serializers/api/admin/product_simple_serializer.rb
@@ -10,16 +10,6 @@ module Api
       def producer_id
         object.supplier_id
       end
-
-      def on_hand
-        return 0 if object.on_hand.nil?
-
-        object.on_hand
-      end
-
-      def price
-        object.price.nil? ? '0.0' : object.price
-      end
     end
   end
 end