#### What? Why?

Closes #5181 

Improves performance on admin inventory page.

#### What should we test?
<!-- List which features should be tested and how. -->

Generally; we need to test that the functionality on the Admin Inventory page is working correctly and there are no regressions in the basic behavior.

More specifically, we're looking at improving the performance of the additional requests that are triggered when the page loads, which fetch the products that the user is able to add to inventory. If you want to test this bit, and you'll need a staging server with lots of data, and a login that has access to a hub with permissions to add lots of producer's products to their inventory. If you load the page with the network tab open, you'll be able to see the requests as they're loading. They'll look like this:

![Screenshot from 2020-04-09 10-36-11](https://user-images.githubusercontent.com/9029026/78875356-0d748800-7a4e-11ea-9b6e-a3206f5cd13d.png)

Another hot tip if you really want to get into it: if you click on one of those requests and then click the **Timing** tab, you should see an indication of how long the request took. Here's two examples showing the response time of the request in the bottom right:

With `master`, 6.20 seconds:
![Screenshot from 2020-04-09 11-25-30](https://user-images.githubusercontent.com/9029026/78880201-554add80-7a55-11ea-8e42-9b11a625eb63.png)

With this PR, 469 milliseconds:
![Screenshot from 2020-04-09 11-25-47](https://user-images.githubusercontent.com/9029026/78880204-5714a100-7a55-11ea-8808-8834e2bc0f3b.png)

Note: I don't recommend testing this as **superadmin** on `master` branch :warning:

#### Release notes
<!-- Write a line or two to be included in the release notes.
Everything is worth mentioning, because you did it for a reason. -->

Improved performance on admin inventory page

<!-- Please assign one category to your PR and delete the others. 
The categories are based on https://keepachangelog.com/en/1.0.0/. -->

Changelog Category: Fixed