@Matt-Yorkley 2020-04-07
I've done a quick round of performance improvements here, the results are good so far. Looking at the overridable products endpoint that is called when this page loads:

**BEFORE:**
- total page load is around **12 seconds**
- the endpoint makes **1100+ queries** (this can be hit many times in one page load)
- the endpoint takes **~6 seconds** to load

![Screenshot from 2020-04-07 15-24-50](https://user-images.githubusercontent.com/9029026/78674863-89eb5780-78e4-11ea-8110-9f43ba85c826.png)

**AFTER:**
- total page load is around **1.2 seconds** (~90% less)
- the endpoint makes **23 queries** (~98% less)
- the endpoint takes **~0.6 seconds** to load (~90% less)

![Screenshot from 2020-04-07 15-22-16](https://user-images.githubusercontent.com/9029026/78674849-8657d080-78e4-11ea-9e22-77857830ec91.png)