@Matt-Yorkley 2020-04-07
The next issue is that the page spams the endpoint with many requests at once, for example with 8 pages it makes 8 requests all at once. Now that the performance is 10 times better, I'm wondering if we can either:
- just have one large request that returns all overridable products
- make sure the requests wait until the last one has loaded before making the next request
- just render all the overridable products using AMS on page load