@sauloperez 2020-04-07
> Or maybe we just go with this for now and see how it looks?

we need to deliver as many improvements as we can. I'd say, if this works, let's get it in and see if it can get better, but at least we've got this.