@Matt-Yorkley 2020-04-08
Thanks @luisramos0! :heart: I'll make those changes in the morning.

> the user will have to wait 50secs until they can use the page? 

I actually found the total time was not so different. I see it locally but it's clear in production as well: the server can't handle being hit with all the requests at once, processes them very slowly and basically in a linear way, but under a huge amount of stress. The endpoint is _incredibly_ slow and CPU intensive (in master) and can take a _very_ long time for the server to actually return all those requests, even when they are all sent at once. 

> Is the user aware that the data is not completely loaded? what happens if the user submits the form when the later pages are still loading, will it work?

This is exactly the problem we have already with this page (in master); the user is not aware it is still loading data, and if they submit things it doesn't necessarily work. It's the cause of: https://github.com/openfoodfoundation/openfoodnetwork/issues/5131

Improving the response time from 6 seconds to 0.6 seconds will help :+1:

Doing a big pagination overhaul here would be great, but I don't think it will be a small job.