@Matt-Yorkley 2020-04-09
This PR certainly doesn't make inventory perfect, it just aims at making it 10x less awful as quickly as possible.