@Matt-Yorkley 2020-04-09
> what do you think about my idea of using the lastPageComplete callback?
I think it could improve UX a lot for slower loads. Could we create an issue to implement it or do you think this PR will be enough?

We can see how things look after this PR. The issue with data not finishing loading is pre-existing. It's possible it will be good enough after this, or maybe we will need to add some additional UI feedback, but it'll definitely need careful inception, dev, and testing if we go for that.