@Matt-Yorkley 2020-04-09
And here's the volume of json data returned per request (same data, different branches).

**BEFORE:**
![Screenshot from 2020-04-09 10-47-00 (copy)](https://user-images.githubusercontent.com/9029026/78876678-fc2c7b00-7a4f-11ea-8f5d-87304587dc62.png)

**AFTER:**
![Screenshot from 2020-04-09 10-47-00](https://user-images.githubusercontent.com/9029026/78876677-fafb4e00-7a4f-11ea-93e4-c49627f93874.png)