@filipefurtad0 2020-04-09
Hi @Matt-Yorkley ,

Whoa... Impressive numbers below!
Page load times, from staging-FR, Fredo's Farm Hub 1311 products.
(left, master /// right, this PR)

![image](https://user-images.githubusercontent.com/49817236/78890332-597af900-7a5d-11ea-8f60-49efc8324923.png)

Also tested the basic functionality of inventory, product review, add or hide, stock override functions. These are also visible in the shopfront, for a regular user.

Moving to ready to go :tada: