@luisramos0 2020-04-09
40secs to 0.4 secs :rofl: :rocket: