@Matt-Yorkley 2020-04-09
I'm not sure if the timing numbers on the right are adjusted for the waiting time of all the (badly queued) requests or something...?

Anyway, at the bottom we have: 

Finish: **1:06 min** -> Finish: **12.44 s**

Fuck yeah.

Thanks again for the awesome testing @filipefurtad0!  :heart: